Decrease modem carrier wait timeout.
When installing a Millennium terminal, the terminal will do an
answer-supervision test by calling into the manager.  With the
default modem carrier wait timeout of 50 seconds, this means the
installer had to wait 50 seconds before starting the download.

Decrease the carrier wait timeout to 3 seconds to speed up the
install process.