mm_manager: Always set the call-in time to AM.
Interestingly, the terminal will call in starting at the call-in time, and continue
calling in at intervals specified, up until midnight.  After that, the terminal will
not call in until the call-in time the following day.  Since we want to call in twice
a day, set the call-in hour to a time in the AM, so the subsequent call 12 hours later
will be in the PM of the same day.